requirements.txt
Flask
python-dotenv
numpy
requests
opencv-python
waitress